Moved SharedArray in parallel.py
